[Booking] Restrict access to update and delete (#7)
* feat: update parameter in get booking controller

* feat: restrict update booking to the issuer only

* feat: restrict delete booking to the issuer only

+++ src/main/java/com/example/eventmanagement/service/impl/BookingServiceImpl.java
@@ -84,8 +84,12 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public BookingEntity updateBookingStatus(Long bookingId, String status) {
+        String email = ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
         Optional<BookingEntity> booking = bookingRepository.findById(bookingId);
         if (booking.isPresent()) {
+            if (!booking.get().getUser().getEmail().equals(email)) {
+                throw new IllegalArgumentException("Unauthorized access");
+            }
             booking.get().setStatus(BookingEntity.BookingStatus.valueOf(status));
             return bookingRepository.save(booking.get());
         }
@@ -94,11 +98,14 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public void deleteBooking(Long bookingId) {
+        String email = ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
         Optional<BookingEntity> booking = bookingRepository.findById(bookingId);
         if (booking.isPresent()) {
+            if (!booking.get().getUser().getEmail().equals(email)) {
+                throw new IllegalArgumentException("Unauthorized access");
+            }
             bookingRepository.delete(booking.get());
-        } else {
-            throw new IllegalArgumentException("Booking not found");
         }
+        throw new IllegalArgumentException("Booking not found");
     }
 }
